+++ CHANGELOG.md
@@ -7,6 +7,8 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 ## [Unreleased]
 
+## [4.2.1] - 2026-06-07
+
 ### Added
 
 - **v4.2.0 cross-repo dogfood evidence doc**: added